small refactoring / restructuring
